AdvancedAPIChallenge/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
Markdown==3.5.1
packaging==23.2
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.4
pytest-django==4.7.0
pytz==2023.3.post1
sqlparse==0.4.4
